added auth2 funtion
