+++ auth.js
@@ -2,9 +2,9 @@ import {
   auth,
   createUserWithEmailAndPassword,
   signInWithEmailAndPassword,
-} from "./firebase/firebase.js";
+} from "./firebase.js";
 
-let loginform = document.getElementById("loginform");
+
 let signupform = document.getElementById("signupform");
 
 
@@ -26,7 +26,7 @@ signupform.addEventListener("submit", (e) => {
     console.log("User signed up:", userCredential.user);
     alert("Signup successful! Redirecting to login...");
 
-    window.location.href = "login.html";
+    window.location.href = "./login.html";
   })
   .catch((error) => {
     console.error("Error:", error.message);
@@ -36,26 +36,7 @@ signupform.addEventListener("submit", (e) => {
 });
 
 
-loginform.addEventListener("submit", (e) => {
-    e.preventDefault();
-    let email = e.target[0].value;
-    let password = e.target[1].value;
 
-    signInWithEmailAndPassword(auth, email, password)
-    .then((userCredential) => {
-        
-        if (userCredential.user.accessToken) {
-            alert("Login successful!");
-
-            
-            window.location.href = "../music.html";
-        }
-    })
-    .catch((error) => {
-        console.error("Error:", error.message);
-        alert(error.message);
-    });
-});
