+++ signup.html
@@ -6,6 +6,7 @@
     <title>signup</title>
     <link rel="stylesheet" href="./login-singup.css">
     <link rel="icon" href="https://storage.googleapis.com/pr-newsroom-wp/1/2023/05/Spotify_Primary_Logo_RGB_Green.png" type="image/svg+xml" />
+    <script src="auth.js"  type="module" defer></script>
     
 </head>
 <body>
@@ -24,6 +25,6 @@
     </div>
 
     
-    <script src="auth.js"  type="module"></script>
+    
 </body>
 </html>
